+++ Changelog.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Cambiado
 - Añadida opcion para mencionar en comando undertale
+- Cambiado praw por asyncpraw
 
 ### Mejorado 
 - Mensajes de error